+++ lib/activerecord-pg-format-db-structure.rb
@@ -10,12 +10,14 @@ require_relative "activerecord-pg-format-db-structure/transforms/inline_foreign_
 require_relative "activerecord-pg-format-db-structure/transforms/move_indices_after_create_table"
 require_relative "activerecord-pg-format-db-structure/transforms/inline_constraints"
 require_relative "activerecord-pg-format-db-structure/transforms/group_alter_table_statements"
+require_relative "activerecord-pg-format-db-structure/transforms/remove_defaults_set_commands"
 require_relative "activerecord-pg-format-db-structure/transforms/sort_schema_migrations"
 require_relative "activerecord-pg-format-db-structure/transforms/sort_table_columns"
 
 module ActiveRecordPgFormatDbStructure
   DEFAULT_TRANSFORMS = [
     Transforms::RemoveCommentsOnExtensions,
+    Transforms::RemoveDefaultsSetCommands,
     Transforms::SortSchemaMigrations,
     Transforms::InlinePrimaryKeys,
     # Transforms::InlineForeignKeys,