+++ spec/activerecord-pg-format-db-structure/formatter_spec.rb
@@ -6,6 +6,16 @@ RSpec.describe ActiveRecordPgFormatDbStructure::Formatter do
       formatter = described_class.new
 
       source = +<<~SQL
+        SET statement_timeout = 0;
+        SET lock_timeout = 0;
+        SET idle_in_transaction_session_timeout = 0;
+        SET client_encoding = 'UTF8';
+        SET standard_conforming_strings = on;
+        SELECT pg_catalog.set_config('search_path', '', false);
+        SET check_function_bodies = true;
+        SET xmloption = content;
+        SET client_min_messages = warning;
+        SET row_security = off;
 
         --
         -- Name: pgcrypto; Type: EXTENSION; Schema: -; Owner: -
@@ -170,6 +180,16 @@ RSpec.describe ActiveRecordPgFormatDbStructure::Formatter do
 
       expect(formatter.format(source)).to eq(<<~SQL.chomp)
 
+
+        SET client_encoding TO "UTF8";
+
+        SELECT pg_catalog.set_config('search_path', '', false);
+
+        SET check_function_bodies TO TRUE;
+
+        SET client_min_messages TO warning;
+
+        SET row_security TO OFF;
 
         CREATE EXTENSION IF NOT EXISTS pgcrypto SCHEMA public;